certificate-service/requirements.txt
Flask==3.0.3
reportlab==4.2.2
python-dotenv==1.0.1
cloudinary==1.40.0